+++ src/main/org/firebirdsql/gds/impl/GDSServerVersion.java
@@ -25,6 +25,7 @@
 package org.firebirdsql.gds.impl;
 
 import java.io.Serializable;
+import java.util.Arrays;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -33,23 +34,27 @@ import java.util.regex.Pattern;
  * in response to the <code>isc_info_firebird_version</code> information call.
  * Expected version format is:
  * <p>
- * <platform>-<type><majorVersion>.<minorVersion>.<variant>.<buildNum> <serverName>[,<extended server info>]
+ * <code>&lt;platform&gt;-&lt;type&gt;&lt;majorVersion&gt;.&lt;minorVersion&gt;.&lt;variant&gt;.&lt;buildNum&gt;
+ * &lt;serverName&gt;</code>, and additional version string elements if present.
  * </p>
  * where <code>platform</code> is a two-character platform identification string,
  * Windows for example is "WI", <code>type</code> is one of the three characters:
  * "V" - production version, "T" - beta version, "X" - development version.
  *
  */
-public class GDSServerVersion implements Serializable {
+public final class GDSServerVersion implements Serializable {
 
 	private static final long serialVersionUID = -153657557318248541L;
 
-	public static final String TYPE_PRODUCTION = "V";
+	@SuppressWarnings("unused")
+    public static final String TYPE_PRODUCTION = "V";
+    @SuppressWarnings("unused")
     public static final String TYPE_BETA = "T";
+    @SuppressWarnings("unused")
     public static final String TYPE_DEVELOPMENT = "X";
 
     private static final Pattern VERSION_PATTERN =
-            Pattern.compile("((\\w{2})-(\\w)(\\d+)\\.(\\d+)\\.(\\d+)\\.(\\d+)(?:-\\S+)?) ([^-,]+)(?:[-,](.*))?");
+            Pattern.compile("((\\w{2})-(\\w)(\\d+)\\.(\\d+)\\.(\\d+)\\.(\\d+)(?:-\\S+)?) (.+)");
 
     private static final int FULL_VERSION_IDX = 1;
     private static final int PLATFORM_IDX = 2;
@@ -59,23 +64,31 @@ public class GDSServerVersion implements Serializable {
     private static final int VARIANT_IDX = 6;
     private static final int BUILD_IDX = 7;
     private static final int SERVER_NAME_IDX = 8;
-    private static final int EXTENDED_INFO_IDX = 9;
 
-    private String rawStr;
+    private final String[] rawVersions;
 
-    private String platform;
-    private String type;
+    private final String platform;
+    private final String type;
 
-    private String fullVersion;
-    private int majorVersion;
-    private int minorVersion;
-    private int variant;
-    private int buildNumber;
+    private final String fullVersion;
+    private final int majorVersion;
+    private final int minorVersion;
+    private final int variant;
+    private final int buildNumber;
 
     private String serverName;
-    private String extendedServerName;
 
-    private GDSServerVersion() {
+    private GDSServerVersion(String[] rawVersions, String platform, String type, String fullVersion, int majorVersion,
+            int minorVersion, int variant, int buildNumber, String serverName) {
+        this.rawVersions = rawVersions.clone();
+        this.platform = platform;
+        this.type = type;
+        this.fullVersion = fullVersion;
+        this.majorVersion = majorVersion;
+        this.minorVersion = minorVersion;
+        this.variant = variant;
+        this.buildNumber = buildNumber;
+        this.serverName = serverName;
     }
 
     public int getBuildNumber() {
@@ -107,7 +120,20 @@ public class GDSServerVersion implements Serializable {
     }
 
     public String getExtendedServerName() {
-        return extendedServerName;
+        if (rawVersions.length < 2) {
+            return null;
+        } else if (rawVersions.length == 2) {
+            return rawVersions[1];
+        } else {
+            StringBuilder sb = new StringBuilder();
+            for (int idx = 1; idx < rawVersions.length; idx++) {
+                if (idx > 1) {
+                    sb.append(',');
+                }
+                sb.append(rawVersions[idx]);
+            }
+            return sb.toString();
+        }
     }
 
     public String getFullVersion() {
@@ -115,7 +141,7 @@ public class GDSServerVersion implements Serializable {
     }
 
     public int hashCode() {
-        return rawStr.hashCode();
+        return Arrays.hashCode(rawVersions);
     }
 
     public boolean equals(Object obj) {
@@ -125,42 +151,52 @@ public class GDSServerVersion implements Serializable {
 
         GDSServerVersion that = (GDSServerVersion) obj;
 
-        return rawStr.equals(that.rawStr);
+        return Arrays.equals(this.rawVersions, that.rawVersions);
     }
 
     public String toString() {
-        return rawStr;
+        if (rawVersions.length == 1) {
+            return rawVersions[0];
+        }
+        StringBuilder sb = new StringBuilder();
+        int idx = 0;
+        sb.append(rawVersions[idx++]);
+        do {
+            sb.append(',');
+            sb.append(rawVersions[idx++]);
+        } while (idx < rawVersions.length);
+        return sb.toString();
     }
 
     /**
      * Parse the raw version string and create a GDSServerVersion object.
      *
-     * @param versionString string to parse.
-     *
-     * @throws GDSServerVersionException if versionString does not match expected pattern
+     * @param versionStrings
+     *         strings to parse, expects a non-empty array with at least 1, and usually 2 version strings
+     * @throws GDSServerVersionException
+     *         if versionString does not match expected pattern
      */
-    public static GDSServerVersion parseRawVersion(String versionString) throws GDSServerVersionException {
-    	Matcher matcher = VERSION_PATTERN.matcher(versionString);
-    	if (!matcher.matches()) {
-            throw new GDSServerVersionException(String.format("Version string \"%s\" does not match expected format",
-                    versionString));
-    	}
+    public static GDSServerVersion parseRawVersion(String... versionStrings) throws GDSServerVersionException {
+        if (versionStrings == null || versionStrings.length == 0 || versionStrings[0] == null) {
+            throw new GDSServerVersionException("No version string information present");
+        }
 
-    	GDSServerVersion version = new GDSServerVersion();
+        Matcher matcher = VERSION_PATTERN.matcher(versionStrings[0]);
+        if (!matcher.matches()) {
+            throw new GDSServerVersionException(
+                    String.format("Version string \"%s\" does not match expected format", versionStrings[0]));
+        }
 
-    	version.rawStr = versionString;
-
-    	version.serverName = matcher.group(SERVER_NAME_IDX);
-    	version.extendedServerName = matcher.group(EXTENDED_INFO_IDX);
-    	version.platform = matcher.group(PLATFORM_IDX);
-    	version.type = matcher.group(TYPE_IDX);
-
-    	version.fullVersion = matcher.group(FULL_VERSION_IDX);
-    	version.majorVersion = Integer.parseInt(matcher.group(MAJOR_IDX));
-    	version.minorVersion = Integer.parseInt(matcher.group(MINOR_IDX));
-    	version.variant = Integer.parseInt(matcher.group(VARIANT_IDX));
-    	version.buildNumber = Integer.parseInt(matcher.group(BUILD_IDX));
-
-    	return version;
+        return new GDSServerVersion(
+                versionStrings,
+                matcher.group(PLATFORM_IDX),
+                matcher.group(TYPE_IDX),
+                matcher.group(FULL_VERSION_IDX),
+                Integer.parseInt(matcher.group(MAJOR_IDX)),
+                Integer.parseInt(matcher.group(MINOR_IDX)),
+                Integer.parseInt(matcher.group(VARIANT_IDX)),
+                Integer.parseInt(matcher.group(BUILD_IDX)),
+                matcher.group(SERVER_NAME_IDX));
     }
+    
 }