+++ src/main/org/firebirdsql/gds/impl/wire/AbstractJavaGDSImpl.java
@@ -34,6 +34,7 @@ import java.net.UnknownHostException;
 
 import java.security.AccessController;
 import java.security.PrivilegedAction;
+import java.util.Arrays;
 import java.util.Map;
 import java.util.HashMap;
 
@@ -414,8 +415,6 @@ public abstract class AbstractJavaGDSImpl extends AbstractGDS implements GDS {
 			ISCConstants.isc_info_firebird_version,
 			ISCConstants.isc_info_ods_version,
 			ISCConstants.isc_info_ods_minor_version,
-			ISCConstants.isc_info_implementation,
-			ISCConstants.isc_info_db_class, ISCConstants.isc_info_base_level,
 			ISCConstants.isc_info_end };
 
 	protected void internalAttachDatabase(DbAttachInfo dbai, IscDbHandle db_handle,
@@ -643,38 +642,22 @@ public abstract class AbstractJavaGDSImpl extends AbstractGDS implements GDS {
 				if (debug)
 					log.debug("isc_info_ods_minor_version:" + value);
 				break;
-			case ISCConstants.isc_info_firebird_version:
+			case ISCConstants.isc_info_firebird_version: {
 				len = iscVaxInteger(info, i, 2);
 				i += 2;
-				byte[] fb_vers = new byte[len - 2];
-				System.arraycopy(info, i + 2, fb_vers, 0, len - 2);
-				i += len;
-				String fb_versS = new String(fb_vers);
-				db.setVersion(fb_versS);
-				if (debug)
-					log.debug("isc_info_firebird_version:" + fb_versS);
+				final int expectedIndex = i + len;
+				final int versionCount = info[i++] & 0xFF;
+				final String[] versionParts = new String[versionCount];
+				for (int versionIndex = 0; versionIndex < versionCount; versionIndex++) {
+					int versionLength = info[i++] & 0xFF;
+					versionParts[versionIndex] = new String(info, i, versionLength);
+					i += versionLength;
+				}
+				assert i == expectedIndex : "Parsing version information lead to wrong index";
+				db.setVersion(versionParts);
+				if (debug) log.debug("isc_info_firebird_version: " + Arrays.toString(versionParts));
 				break;
-			case ISCConstants.isc_info_implementation:
-				len = iscVaxInteger(info, i, 2);
-				i += 2;
-				byte[] impl = new byte[len - 2];
-				System.arraycopy(info, i + 2, impl, 0, len - 2);
-				i += len;
-				break;
-			case ISCConstants.isc_info_db_class:
-				len = iscVaxInteger(info, i, 2);
-				i += 2;
-				byte[] db_class = new byte[len - 2];
-				System.arraycopy(info, i + 2, db_class, 0, len - 2);
-				i += len;
-				break;
-			case ISCConstants.isc_info_base_level:
-				len = iscVaxInteger(info, i, 2);
-				i += 2;
-				byte[] base_level = new byte[len - 2];
-				System.arraycopy(info, i + 2, base_level, 0, len - 2);
-				i += len;
-				break;
+			}
 			case ISCConstants.isc_info_truncated:
 				if (debug)
 					log.debug("isc_info_truncated ");